requirements.txt
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-mock==3.15.1
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
sentry-sdk==2.51.0
shellingham==1.5.4
SQLAlchemy==2.0.46
sqlmodel==0.0.32
starlette==0.50.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
